requirements.txt

# LangChain core
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-huggingface==0.0.3
#langchain-mistralai>=0.1.0

transformers==4.44.2
huggingface_hub==0.24.6


# API & server
fastapi>=0.110.0
uvicorn>=0.29.0

# Vector store / embeddings
faiss-cpu>=1.7.4

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tiktoken>=0.7.0
pydantic>=2.0.0